Update fetch page size, refactor background refresh sagas
As first steps to help mitigate performance issues on the
ovirt-engine's DB due to web-ui polling expensive queries
at a high frequency, a few change have been made.

  - The default VM and Pool fetch page size is now 20, closer
    to a reasonable non-Admin user number of VM.  This can help
    reduce the number of calls made on login.

  - Refactored the background data refresh sagas to
    `background-refresh.js` to keep them logically isolated.

  - The refresh scheduler is now controlled by the `changePage()`
    saga instead of the `login()` saga.  This will help by keeping
    a consistent time between when data is initially fetched for a
    page and when the refresh will happen.

  - The refresh now throttles calls to run only 1 refresh at the
    start and end of a 5 second interval (instead of a 1 second
    interval). This helps prevent a manual refresh from running
    too often or running at the same time as a scheduled refresh.
