+++ src/sagas/index.js
@@ -2,8 +2,6 @@ import {
   all,
   call,
   put,
-  race,
-  take,
   takeEvery,
   takeLatest,
   throttle,
@@ -12,14 +10,14 @@ import {
 import { push } from 'connected-react-router'
 
 import Api, { Transforms } from '_/ovirtapi'
-import AppConfiguration from '_/config'
 import { saveToLocalStorage } from '_/storage'
 
+import sagasRefresh from './background-refresh'
 import sagasDisks from './disks'
 import sagasLogin from './login'
 import sagasOptionsDialog from '_/components/OptionsDialog/sagas'
 import sagasRoles from './roles'
-import sagasStorageDomains, { fetchIsoFiles } from './storageDomains'
+import sagasStorageDomains from './storageDomains'
 import sagasVmChanges from './vmChanges'
 import sagasVmSnapshots from '_/components/VmDetails/cards/SnapshotsCard/sagas'
 
@@ -29,32 +27,22 @@ import {
   updateVms,
   removeVms,
   vmActionInProgress,
-  removeMissingVms,
 
   getSingleVm,
-  selectVmDetail as actionSelectVmDetail,
   setVmSnapshots,
 
   setUserMessages,
   dismissUserMessage,
 
   getSinglePool,
-  removeMissingPools,
   removePools,
   updatePools,
   updateVmsPoolsCount,
   poolActionInProgress,
 
-  refresh,
-  getVmsByCount,
-  getPoolsByCount,
-  getIsoFiles,
-  getConsoleOptions as actionGetConsoleOptions,
   setVmNics,
   removeActiveRequest,
-  stopSchedulerFixedDelay,
   getVmCdRom,
-  setCurrentPage,
   setVmsFilters,
 } from '_/actions'
 
@@ -80,7 +68,6 @@ import {
 
 import {
   ADD_VM_NIC,
-  CHANGE_PAGE,
   OPEN_CONSOLE_MODAL,
   CHECK_TOKEN_EXPIRED,
   CLEAR_USER_MSGS,
@@ -97,20 +84,11 @@ import {
   GET_RDP_VM,
   GET_VMS_BY_COUNT,
   GET_VMS_BY_PAGE,
-  REFRESH_DATA,
   SAVE_CONSOLE_OPTIONS,
   SAVE_FILTERS,
   SELECT_POOL_DETAIL,
   SELECT_VM_DETAIL,
   NAVIGATE_TO_VM_DETAILS,
-  START_SCHEDULER_FIXED_DELAY,
-  STOP_SCHEDULER_FIXED_DELAY,
-
-  CONSOLE_PAGE_TYPE,
-  DETAIL_PAGE_TYPE,
-  DIALOG_PAGE_TYPE,
-  MAIN_PAGE_TYPE,
-  NO_REFRESH_TYPE,
 } from '_/constants'
 
 import {
@@ -119,7 +97,6 @@ import {
   canUserEditVm,
   canUserEditVmStorage,
   canUserManipulateSnapshots,
-  isNumber,
 } from '_/utils'
 
 const vmFetchAdditionalList =
@@ -136,123 +113,6 @@ const vmFetchAdditionalList =
 
 export const EVERYONE_GROUP_ID = 'eee00000-0000-0000-0000-123456789eee'
 
-function* fetchByPage (action) {
-  yield put(setChanged({ value: false }))
-  yield fetchVmsByPage(action)
-  yield fetchPoolsByPage(action)
-}
-
-function* getIdsByType (type) {
-  const ids = yield select(state =>
-    state.vms
-      .get(type)
-      .reduce((entityIds, entity, entityId) => {
-        entityIds.push(entityId)
-        return entityIds
-      }, [])
-  )
-  return ids
-}
-
-function* refreshMainPage ({ shallowFetch, page }) {
-  shallowFetch = !!shallowFetch
-
-  // refresh VMs and remove any that haven't been refreshed
-  const fetchedVmIds = yield fetchVmsByCount(getVmsByCount({
-    count: page * AppConfiguration.pageLimit,
-    shallowFetch,
-  }))
-
-  const vmsIds = yield getIdsByType('vms')
-  const fetchedDirectlyVmIds =
-    (yield all(
-      vmsIds
-        .filter(vmId => !fetchedVmIds.includes(vmId))
-        .map(vmId => call(fetchSingleVm, getSingleVm({ vmId, shallowFetch })))
-    ))
-      .reduce((vmIds, vm) => { if (vm) vmIds.push(vm.id); return vmIds }, [])
-
-  yield put(removeMissingVms({ vmIdsToPreserve: [ ...fetchedVmIds, ...fetchedDirectlyVmIds ] }))
-
-  // refresh Pools and remove any that haven't been refreshed
-  const fetchedPoolIds = yield fetchPoolsByCount(getPoolsByCount({
-    count: page * AppConfiguration.pageLimit,
-  }))
-  const filteredPoolIds = yield getIdsByType('pools')
-  const fetchedDirectlyPoolIds =
-    (yield all(
-      filteredPoolIds
-        .filter(poolId => !fetchedPoolIds.includes(poolId))
-        .map(poolId => call(fetchSinglePool, getSinglePool({ poolId })))
-    ))
-      .reduce((poolIds, pool) => { if (pool) poolIds.push(pool.id); return poolIds }, [])
-
-  yield put(removeMissingPools({ poolIdsToPreserve: [ ...fetchedPoolIds, ...fetchedDirectlyPoolIds ] }))
-
-  // update counts
-  yield put(updateVmsPoolsCount())
-}
-
-function* refreshDetailPage ({ id, onNavigation, onSchedule }) {
-  yield selectVmDetail(actionSelectVmDetail({ vmId: id }))
-  yield getConsoleOptions(actionGetConsoleOptions({ vmId: id }))
-
-  // Load ISO images on manual refresh click only
-  if (!onNavigation && !onSchedule) {
-    yield fetchIsoFiles(getIsoFiles())
-  }
-}
-
-function* refreshDialogPage ({ id, onNavigation, onSchedule }) {
-  if (id) {
-    yield selectVmDetail(actionSelectVmDetail({ vmId: id }))
-  }
-
-  // Load ISO images on manual refresh click only
-  if (!onNavigation && !onSchedule) {
-    yield fetchIsoFiles(getIsoFiles())
-  }
-}
-
-function* refreshConsolePage ({ id }) {
-  if (id) {
-    yield selectVmDetail(actionSelectVmDetail({ vmId: id }))
-  }
-}
-
-const pagesRefreshers = {
-  [MAIN_PAGE_TYPE]: refreshMainPage,
-  [DETAIL_PAGE_TYPE]: refreshDetailPage,
-  [DIALOG_PAGE_TYPE]: refreshDialogPage,
-  [CONSOLE_PAGE_TYPE]: refreshConsolePage,
-}
-
-function* refreshData (action) {
-  const currentPage = yield select(state => state.config.get('currentPage'))
-  const refreshType =
-    currentPage.type === NO_REFRESH_TYPE ? null
-      : currentPage.type === undefined ? MAIN_PAGE_TYPE
-        : currentPage.type
-
-  console.info('refreshData() 🡒 payload:', action.payload, 'currentPage:', currentPage, 'refreshType:', refreshType)
-  if (refreshType) {
-    yield pagesRefreshers[refreshType](Object.assign({ id: currentPage.id }, action.payload))
-  }
-  console.info('refreshData() 🡒 finished')
-}
-
-/**
- * Change the current page type (based on `routes.js`) at the time of navigation to the page.
- */
-function* changePage (action) {
-  yield put(setCurrentPage(action.payload))
-  yield refreshData(refresh({
-    onNavigation: true,
-    shallowFetch: true,
-    page: yield select(state => state.vms.get('page')),
-  }))
-}
-
 export function* transformAndPermitVm (vm) {
   const internalVm = Transforms.VM.toInternal({ vm })
   internalVm.userPermits = yield entityPermissionsToUserPermits(internalVm)
@@ -265,10 +125,17 @@ export function* transformAndPermitVm (vm) {
   return internalVm
 }
 
+export function* fetchByPage (action) {
+  console.info(`fetchByPage, page: ${action.payload.page}`)
+  yield put(setChanged({ value: false }))
+  yield fetchVmsByPage(action)
+  yield fetchPoolsByPage(action)
+}
+
 /**
  * Fetch VMs with additional nested data requested
  */
-function* fetchVmsByPage (action) {
+export function* fetchVmsByPage (action) {
   const { shallowFetch, page } = action.payload
 
   action.payload.additional = shallowFetch ? ['graphics_consoles'] : vmFetchAdditionalList
@@ -292,7 +159,7 @@ function* fetchVmsByPage (action) {
 /**
  * Fetch a given number of VMs (**action.payload.count**).
  */
-function* fetchVmsByCount (action) {
+export function* fetchVmsByCount (action) {
   const { shallowFetch } = action.payload
   const fetchedVmIds = []
 
@@ -306,7 +173,6 @@ function* fetchVmsByCount (action) {
       fetchedVmIds.push(internalVm.id)
       internalVms.push(internalVm)
     }
-    console.log('internalVms:', internalVms)
 
     yield put(updateVms({ vms: internalVms, copySubResources: shallowFetch }))
     yield fetchUnknownIcons({ vms: internalVms })
@@ -370,7 +236,7 @@ export function* fetchSingleVm (action) {
   return internalVm
 }
 
-function* fetchPoolsByCount (action) {
+export function* fetchPoolsByCount (action) {
   const fetchedPoolIds = []
 
   const allPools = yield callExternalAction('getPoolsByCount', Api.getPoolsByCount, action)
@@ -385,7 +251,7 @@ function* fetchPoolsByCount (action) {
   return fetchedPoolIds
 }
 
-function* fetchPoolsByPage (action) {
+export function* fetchPoolsByPage (action) {
   const allPools = yield callExternalAction('getPoolsByPage', Api.getPoolsByPage, action)
 
   if (allPools && allPools['vm_pool']) {
@@ -396,7 +262,7 @@ function* fetchPoolsByPage (action) {
   }
 }
 
-function* fetchSinglePool (action) {
+export function* fetchSinglePool (action) {
   const { poolId } = action.payload
 
   const pool = yield callExternalAction('getPool', Api.getPool, action, true)
@@ -664,59 +530,6 @@ function* delayedRemoveActiveRequest ({ payload: requestId }) {
   yield put(removeActiveRequest(requestId))
 }
 
-function* startSchedulerWithFixedDelay (action) {
-  // if a scheduler is already running, stop it
-  yield put(stopSchedulerFixedDelay())
-
-  // run a new scheduler
-  yield schedulerWithFixedDelay(action.payload.delayInSeconds)
-}
-
-let _SchedulerCount = 0
-
-function* schedulerWithFixedDelay (delayInSeconds = AppConfiguration.schedulerFixedDelayInSeconds) {
-  if (!isNumber(delayInSeconds) || delayInSeconds <= 0) {
-    return
-  }
-
-  const myId = _SchedulerCount++
-  console.log(`⏰ schedulerWithFixedDelay[${myId}] starting fixed delay scheduler`)
-
-  let enabled = true
-  while (enabled) {
-    console.log(`⏰ schedulerWithFixedDelay[${myId}] stoppable delay for: ${delayInSeconds}`)
-    const { stopped } = yield race({
-      stopped: take(STOP_SCHEDULER_FIXED_DELAY),
-      fixedDelay: call(delay, (delayInSeconds * 1000)),
-    })
-
-    const isTokenExpired = yield select(state => state.config.get('isTokenExpired'))
-    if (isTokenExpired) {
-      enabled = false
-      console.log(`⏰ schedulerWithFixedDelay[${myId}] scheduler has been stopped due token expiration`)
-      continue
-    }
-
-    if (stopped) {
-      enabled = false
-      console.log(`⏰ schedulerWithFixedDelay[${myId}] scheduler has been stopped`)
-    } else {
-      console.log(`⏰ schedulerWithFixedDelay[${myId}] running after delay of: ${delayInSeconds}`)
-
-      const oVirtVersion = yield select(state => state.config.get('oVirtApiVersion'))
-      if (oVirtVersion.get('passed')) {
-        yield refreshData(refresh({
-          onSchedule: true,
-          shallowFetch: true,
-          page: yield select(state => state.vms.get('page')),
-        }))
-      } else {
-        console.log(`⏰ schedulerWithFixedDelay[${myId}] event skipped since oVirt API version does not match`)
-      }
-    }
-  }
-}
-
 function* navigateToVmDetails ({ payload: { vmId } }) {
   yield put(push(`/vm/${vmId}`))
 }
@@ -724,13 +537,10 @@ function* navigateToVmDetails ({ payload: { vmId } }) {
 export function* rootSaga () {
   yield all([
     ...sagasLogin,
+    ...sagasRefresh,
+
     takeLatest(CHECK_TOKEN_EXPIRED, doCheckTokenExpired),
     takeEvery(DELAYED_REMOVE_ACTIVE_REQUEST, delayedRemoveActiveRequest),
-
-    takeEvery(START_SCHEDULER_FIXED_DELAY, startSchedulerWithFixedDelay),
-    // STOP_SCHEDULER_FIXED_DELAY is taken by `schedulerWithFixedDelay()`
-    throttle(1000, REFRESH_DATA, refreshData),
-    takeLatest(CHANGE_PAGE, changePage),
 
     throttle(100, GET_BY_PAGE, fetchByPage),
     throttle(100, GET_VMS_BY_COUNT, fetchVmsByCount),
