+++ src/sagas/login.js
@@ -10,7 +10,6 @@ import {
   loginSuccessful,
   loginFailed,
   appConfigured,
-  startSchedulerFixedDelay,
 
   failedExternalAction,
   setOvirtApiVersion,
@@ -106,8 +105,11 @@ function* login (action) {
   yield initialLoad()
   console.groupEnd('Login Data Fetch')
 
+  // The first page of VMs and Pools are loaded by the background-refresh sagas when the
+  // `react-router-config` loads the route '/'.  Loading of additional pages is handled by
+  // user interaction (scrolling) on the `Vms` card view component.
+
   yield put(appConfigured())
-  yield put(startSchedulerFixedDelay())
   yield autoConnectCheck()
   yield put(getAllEvents())
 }
@@ -220,8 +222,7 @@ function* initialLoad () {
   console.log('\u2714 data loads that require storage domains are complete')
   console.groupEnd('needs storage domains')
 
-  // The `Vms` card view component will take care of loading pages of VMs and Pools as needed.
-  // Loading VMs and Pools here is not necessary and will cause issues with `Vms`'s loading.
+  // Vms and Pools are loaded as needed / accessed
 }
 
 function* autoConnectCheck () {
